requirements.txt
fastapi
sqlalchemy
uvicorn[standard]
pyjwt
requests